refactor: channel talk wrapping
