+++ src/lib/provider/appProvier.tsx
@@ -2,13 +2,13 @@
 import { PropsWithChildren } from 'react';
 import { ApolloSetting } from './apollo';
 import RecoilSetting from './recoil';
-import ChannelTalkBoot from '@/features/channeltalk/components/ChannelTalkBoot';
+import CustomerServiceBoot from '../customerservice/CustomerServiceBoot';
 
 const AppProvider = ({ children }: PropsWithChildren) => {
   return (
     <RecoilSetting>
       <ApolloSetting>{children}</ApolloSetting>
-      <ChannelTalkBoot />
+      <CustomerServiceBoot />
     </RecoilSetting>
   );
 };
